requirements.txt
psycopg2-binary
python-dotenv
langchain-community
langchain-openai